homework_6/python_consumer_producer/requirements.txt
kafka-python
faker
